requirements.txt
python-telegram-bot[job-queue]>=21.10,<22
pytest>=8.0
